internal: refactor find_map diagnostic
